refactor: unify SettingHandler — single registry in agent/setting_runtime.go
- Created agent/setting_runtime.go: single SettingHandlerRegistry for both CLI and server.
  All runtime setting effects (config updates + backend side-effects) in one place.
- Deleted cmd/xbot-cli/setting_handlers.go (239 lines) — CLI uses agent.ApplyRuntimeSettings.
- Simplified serverapp/setting_handlers.go to thin wrapper (32 lines) — adds saveServerConfig.
- Updated CLI ApplySettings callback + removed dead globalValues block.
- Updated tests to reference agent.SettingHandlerRegistry.

main.go: 3043 → 2823 lines (-220). All tests pass.
