--- cmd/xbot-cli/setting_handlers.go
@@ -1,201 +1,0 @@
-package main
-
-import (
-	"strconv"
-	"strings"
-
-	"xbot/agent"
-	"xbot/channel"
-	"xbot/config"
-	log "xbot/logger"
-)
-
-// cliSettingHandler defines how a setting key affects CLI-side runtime state.
-// Each field is optional — a key that only updates config leaves ApplyBackend nil.
-type cliSettingHandler struct {
-	// ApplyConfig updates the in-memory config struct. cfg is always non-nil.
-	ApplyConfig func(cfg *config.Config, value string)
-	// ApplyBackend applies runtime side effects via the backend.
-	// backend and senderID are always non-nil/non-empty when called.
-	// chatID is the current session's chat ID (may be empty for global operations).
-	ApplyBackend func(backend agent.AgentBackend, senderID, chatID, value string)
-}
-
-// cliSettingHandlers is the CLI-side counterpart of serverapp.settingHandlerRegistry.
-// Every key in channel.CLIRuntimeSettingKeys that needs CLI-side handling MUST have
-// an entry here. TestCLISettingHandlersCoversAllRuntimeKeys enforces this.
-//
-// To add a new runtime setting:
-//  1. Add the key to channel.CLIRuntimeSettingKeys
-//  2. Add a handler here (and in serverapp/setting_handlers.go)
-//  3. Done.
-var cliSettingHandlers = map[string]cliSettingHandler{
-	// --- LLM tier settings (config-only, LLM rebuild handled by caller) ---
-	"vanguard_model": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.LLM.VanguardModel = strings.TrimSpace(value)
-		},
-	},
-	"balance_model": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.LLM.BalanceModel = strings.TrimSpace(value)
-		},
-	},
-	"swift_model": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.LLM.SwiftModel = strings.TrimSpace(value)
-		},
-	},
-
-	// --- Agent settings ---
-	"sandbox_mode": {
-		ApplyConfig: func(cfg *config.Config, value string) { cfg.Sandbox.Mode = value },
-		// sandbox ReinitSandbox is handled separately in local mode (needs app.workDir)
-	},
-	"compression_threshold": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			if f, err := strconv.ParseFloat(value, 64); err == nil && f > 0 {
-				cfg.Agent.CompressionThreshold = f
-			}
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if f, err := strconv.ParseFloat(value, 64); err == nil && f > 0 {
-				backend.SetCompressionThreshold(f)
-			}
-		},
-	},
-	"memory_provider": {
-		ApplyConfig: func(cfg *config.Config, value string) { cfg.Agent.MemoryProvider = value },
-	},
-	"tavily_api_key": {}, // Stored in user_settings; WebSearchTool reads dynamically via ConfigGet
-
-	// --- Runtime state settings (config + backend side-effects) ---
-	"context_mode": {
-		ApplyConfig: func(cfg *config.Config, value string) { cfg.Agent.ContextMode = value },
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			_ = backend.SetContextMode(value)
-		},
-	},
-	"max_iterations": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.Agent.MaxIterations = channel.ParseSettingInt(value, cfg.Agent.MaxIterations)
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if n, err := strconv.Atoi(value); err == nil && n > 0 {
-				backend.SetMaxIterations(n)
-			}
-		},
-	},
-	"max_concurrency": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.Agent.MaxConcurrency = channel.ParseSettingInt(value, cfg.Agent.MaxConcurrency)
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if n, err := strconv.Atoi(value); err == nil && n > 0 {
-				backend.SetMaxConcurrency(n)
-			}
-		},
-	},
-	"max_context_tokens": {
-		// max_context is subscription-scoped, stored in PerModelConfigs.
-		// Do NOT write to cfg.Agent.MaxContextTokens (global fallback only).
-		ApplyConfig: nil,
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if n, err := strconv.Atoi(value); err == nil && n >= 0 {
-				if chatID != "" {
-					backend.SetMaxContextTokens(n, chatID)
-				} else {
-					backend.SetMaxContextTokens(n)
-				}
-			}
-		},
-	},
-	// enable_auto_compress is a legacy alias for context_mode.
-	// Its ApplyBackend calls SetContextMode, so context_mode must be processed LAST
-	// in batch mode to correctly override when both are present.
-	"enable_auto_compress": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			b := channel.ParseSettingBool(value)
-			cfg.Agent.EnableAutoCompress = &b
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if channel.ParseSettingBool(value) {
-				_ = backend.SetContextMode("auto")
-			} else {
-				_ = backend.SetContextMode("none")
-			}
-		},
-	},
-
-	// --- Layout settings (UI-only, no config/backend side-effects) ---
-	// These keys are marked Runtime=true so the CLI can detect layout changes
-	// and trigger relayout. The actual field updates happen in handleSettingsSavedMsg.
-	"layout_mode":      {},
-	"sidebar_enabled":  {},
-	"sidebar_width":    {},
-	"sidebar_position": {},
-	"sidebar_sections": {},
-	"chat_max_width":   {},
-	"chat_center":      {},
-}
-
-// applyCLISettingsToConfig applies config field updates for all recognized keys in values.
-// Returns the set of keys that were handled (for logging/warning about unhandled keys).
-func applyCLISettingsToConfig(cfg *config.Config, values map[string]string) map[string]bool {
-	handled := make(map[string]bool, len(values))
-	for k, v := range values {
-		h, ok := cliSettingHandlers[k]
-		if !ok {
-			continue
-		}
-		handled[k] = true
-		if h.ApplyConfig != nil {
-			h.ApplyConfig(cfg, v)
-		}
-	}
-	return handled
-}
-
-// applyCLISettingsToBackend applies runtime backend side-effects for all recognized keys.
-// context_mode is processed LAST so it correctly overrides enable_auto_compress when both are present.
-func applyCLISettingsToBackend(backend agent.AgentBackend, senderID, chatID string, values map[string]string) {
-	// Process all keys except context_mode first
-	for k, v := range values {
-		if k == "context_mode" {
-			continue // process last
-		}
-		h, ok := cliSettingHandlers[k]
-		if !ok {
-			if !isKnownNonRuntimeKey(k) {
-				log.WithField("key", k).Warn("applyCLISettings: unhandled setting key")
-			}
-			continue
-		}
-		if h.ApplyBackend != nil {
-			h.ApplyBackend(backend, senderID, chatID, v)
-		}
-	}
-	// Process context_mode last so it overrides enable_auto_compress
-	if v, ok := values["context_mode"]; ok && v != "" {
-		if h, ok := cliSettingHandlers["context_mode"]; ok && h.ApplyBackend != nil {
-			h.ApplyBackend(backend, senderID, chatID, v)
-		}
-	}
-}
-
-// isKnownNonRuntimeKey returns true for keys that don't need runtime handling
-// (UI-only, persistence-only, or action keys).
-func isKnownNonRuntimeKey(key string) bool {
-	// Keys not in AllSettingDefs are known non-runtime — use shared function.
-	isKnown := channel.IsKnownNonRuntimeKey(key)
-	if !isKnown {
-		log.WithField("key", key).Warn("setting key has no CLI handler and is not a known non-runtime key")
-	}
-	return isKnown
-}
-
-// missingCLIHandlerKeys returns keys from channel.CLIRuntimeSettingKeys
-// that are missing from cliSettingHandlers.
-func missingCLIHandlerKeys() []string {
-	return channel.MissingRegistryKeys(cliSettingHandlers)
-}
