+++ serverapp/setting_handlers.go
@@ -1,238 +1,32 @@
 package serverapp
 
 import (
-	"strconv"
-	"strings"
-
 	"xbot/agent"
-	"xbot/channel"
 	"xbot/config"
-	log "xbot/logger"
-	"xbot/tools"
 )
 
-// settingHandler defines how a setting key updates runtime state.
-// Each field is optional — a key that only updates config leaves ApplyBackend nil.
-type settingHandler struct {
-	// ApplyConfig updates the in-memory config struct. cfg is always non-nil.
-	ApplyConfig func(cfg *config.Config, value string)
-	// ApplyBackend applies runtime side effects via the backend.
-	// Called after ApplyConfig. Both backend and senderID are non-nil/non-empty.
-	// chatID is the current session's chat ID (may be empty for global operations).
-	ApplyBackend func(backend agent.AgentBackend, senderID, chatID, value string)
-	// ApplyFull is called with both cfg and backend. Used when the side effect
-	// needs config context (e.g. sandbox reinit needs cfg.Agent.WorkDir).
-	// If set, called instead of the ApplyConfig+ApplyBackend pair.
-	ApplyFull func(cfg *config.Config, backend agent.AgentBackend, senderID, value string)
-}
-
-// settingHandlerRegistry is the single source of truth for server-side runtime
-// setting effects. Every key in channel.CLIRuntimeSettingKeys that needs server-side
-// handling MUST have an entry here.
+// setting_handlers.go — thin wrapper around agent.SettingHandlerRegistry.
+// Serverapp only adds server-specific config persistence (saveServerConfig).
 //
 // To add a new runtime setting:
 //  1. Add the key to channel.CLIRuntimeSettingKeys
-//  2. Add a handler here (and in cmd/xbot-cli/setting_handlers.go for CLI side)
-//  3. Done — no switch-case to update, no if-chain to extend.
-//
-// If a key appears in channel.CLIRuntimeSettingKeys but is missing here,
-// TestAllRuntimeKeysHaveHandlers will fail at CI time.
-var settingHandlerRegistry = map[string]settingHandler{
-	// --- LLM tier settings (config-only, backend effects applied by caller via SetModelTiers) ---
-	"vanguard_model": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.LLM.VanguardModel = strings.TrimSpace(value)
-		},
-	},
-	"balance_model": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.LLM.BalanceModel = strings.TrimSpace(value)
-		},
-	},
-	"swift_model": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.LLM.SwiftModel = strings.TrimSpace(value)
-		},
-	},
+//  2. Add a handler in agent/setting_runtime.go
+//  3. Done.
 
-	// --- Agent settings ---
-	"sandbox_mode": {
-		ApplyConfig: func(cfg *config.Config, value string) { cfg.Sandbox.Mode = value },
-		ApplyFull: func(cfg *config.Config, backend agent.AgentBackend, senderID, value string) {
-			workDir := cfg.Agent.WorkDir
-			if workDir == "" {
-				workDir = "."
-			}
-			tools.ReinitSandbox(cfg.Sandbox, workDir)
-			backend.SetSandbox(tools.GetSandbox(), value)
-		},
-	},
-	"compression_threshold": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			if f, err := strconv.ParseFloat(value, 64); err == nil && f > 0 {
-				cfg.Agent.CompressionThreshold = f
-			}
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if f, err := strconv.ParseFloat(value, 64); err == nil && f > 0 {
-				backend.SetCompressionThreshold(f)
-			}
-		},
-	},
-	"memory_provider": {
-		ApplyConfig: func(cfg *config.Config, value string) { cfg.Agent.MemoryProvider = value },
-	},
-	"tavily_api_key": {}, // Stored in user_settings; WebSearchTool reads dynamically via ConfigGet
-
-	// --- Runtime state settings (config + backend side-effects) ---
-	"context_mode": {
-		ApplyConfig: func(cfg *config.Config, value string) { cfg.Agent.ContextMode = value },
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			_ = backend.SetContextMode(value)
-		},
-	},
-	"max_iterations": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.Agent.MaxIterations = channel.ParseSettingInt(value, cfg.Agent.MaxIterations)
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if n, err := strconv.Atoi(value); err == nil && n > 0 {
-				backend.SetMaxIterations(n)
-			}
-		},
-	},
-	"max_concurrency": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			cfg.Agent.MaxConcurrency = channel.ParseSettingInt(value, cfg.Agent.MaxConcurrency)
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if n, err := strconv.Atoi(value); err == nil && n > 0 {
-				backend.SetMaxConcurrency(n)
-			}
-		},
-	},
-	"max_context_tokens": {
-		// max_context is a subscription-scoped property stored in PerModelConfigs.
-		// Do NOT write to cfg.Agent.MaxContextTokens (that's the global fallback only).
-		ApplyConfig: nil,
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if n, err := strconv.Atoi(value); err == nil && n >= 0 {
-				if chatID != "" {
-					backend.SetMaxContextTokens(n, chatID)
-				} else {
-					backend.SetMaxContextTokens(n)
-				}
-			}
-		},
-	},
-	// enable_auto_compress is a legacy alias for context_mode.
-	// Its ApplyBackend calls SetContextMode, so context_mode must be processed LAST
-	// in batch mode to correctly override when both are present.
-	"enable_auto_compress": {
-		ApplyConfig: func(cfg *config.Config, value string) {
-			b := channel.ParseSettingBool(value)
-			cfg.Agent.EnableAutoCompress = &b
-		},
-		ApplyBackend: func(backend agent.AgentBackend, senderID, chatID, value string) {
-			if channel.ParseSettingBool(value) {
-				_ = backend.SetContextMode("auto")
-			} else {
-				_ = backend.SetContextMode("none")
-			}
-		},
-	},
-
-	// --- Layout settings (UI-only, no server-side side-effects) ---
-	// These keys are marked Runtime=true so the CLI can detect layout changes.
-	// On the server side they're no-ops — the actual layout is handled by the CLI frontend.
-	"layout_mode":      {},
-	"sidebar_enabled":  {},
-	"sidebar_width":    {},
-	"sidebar_position": {},
-	"sidebar_sections": {},
-	"chat_max_width":   {},
-	"chat_center":      {},
-}
-
-// applyRuntimeSetting applies a single setting change to the in-memory config and backend.
-// Used by admin RPC handler after the setting is persisted to DB.
-// For batch operations (startup sync), use applyRuntimeSettings instead.
+// applyRuntimeSetting applies a single setting change and saves server config.
 func applyRuntimeSetting(cfg *config.Config, backend agent.AgentBackend, senderID, key, value string) {
-	handler, ok := settingHandlerRegistry[key]
-	if !ok {
-		if !channel.IsKnownNonRuntimeKey(key) {
-			log.WithField("key", key).WithField("value", value).
-				Warn("applyRuntimeSetting: unhandled setting key, ignoring")
-		}
-		return
-	}
-	if handler.ApplyFull != nil && backend != nil {
-		handler.ApplyFull(cfg, backend, senderID, value)
-	} else {
-		if handler.ApplyConfig != nil {
-			handler.ApplyConfig(cfg, value)
-		}
-		if handler.ApplyBackend != nil && backend != nil {
-			handler.ApplyBackend(backend, senderID, "", value)
-		}
-	}
-	if backend != nil && backend.LLMFactory() != nil {
-		backend.LLMFactory().SetModelTiers(cfg.LLM)
-	}
+	agent.ApplyRuntimeSetting(cfg, backend, senderID, key, value)
 	_ = saveServerConfig(cfg)
 }
 
-// applyRuntimeSettings applies a batch of setting changes to the in-memory config and backend.
-// context_mode is processed LAST so it correctly overrides enable_auto_compress when both are present.
-// SetModelTiers and saveServerConfig are called once after all keys are processed.
+// applyRuntimeSettings applies a batch of setting changes and saves server config.
 func applyRuntimeSettings(cfg *config.Config, backend agent.AgentBackend, senderID string, values map[string]string) {
-	// Process all keys except context_mode first
-	for k, v := range values {
-		if k == "context_mode" {
-			continue // process last
-		}
-		handler, ok := settingHandlerRegistry[k]
-		if !ok {
-			if !channel.IsKnownNonRuntimeKey(k) {
-				log.WithField("key", k).WithField("value", v).
-					Warn("applyRuntimeSettings: unhandled setting key, ignoring")
-			}
-			continue
-		}
-		if handler.ApplyFull != nil && backend != nil {
-			handler.ApplyFull(cfg, backend, senderID, v)
-		} else {
-			if handler.ApplyConfig != nil {
-				handler.ApplyConfig(cfg, v)
-			}
-			if handler.ApplyBackend != nil && backend != nil {
-				handler.ApplyBackend(backend, senderID, "", v)
-			}
-		}
-	}
-	// Process context_mode last so it overrides enable_auto_compress
-	if v, ok := values["context_mode"]; ok && v != "" {
-		handler := settingHandlerRegistry["context_mode"]
-		if handler.ApplyFull != nil && backend != nil {
-			handler.ApplyFull(cfg, backend, senderID, v)
-		} else {
-			if handler.ApplyConfig != nil {
-				handler.ApplyConfig(cfg, v)
-			}
-			if handler.ApplyBackend != nil && backend != nil {
-				handler.ApplyBackend(backend, senderID, "", v)
-			}
-		}
-	}
-	// SetModelTiers and config save: once after all keys
-	if backend != nil && backend.LLMFactory() != nil {
-		backend.LLMFactory().SetModelTiers(cfg.LLM)
-	}
+	agent.ApplyRuntimeSettings(cfg, backend, senderID, values)
 	_ = saveServerConfig(cfg)
 }
 
 // missingHandlerKeys returns keys from channel.CLIRuntimeSettingKeys
-// that are missing from settingHandlerRegistry.
+// that are missing from agent.SettingHandlerRegistry.
 func missingHandlerKeys() []string {
-	return channel.MissingRegistryKeys(settingHandlerRegistry)
+	return agent.MissingSettingHandlerKeys()
 }
